updated ReadData to modular file structure
